1-news-metadata/requirements.txt
langgraph==1.0.1
langchain==1.0.2
langchain-openai==1.0.1
python-dotenv==1.1.1